refactor: move ErrorBoundary + OnboardingTour to shared/
